+++ docs/changes.rst
@@ -7,6 +7,7 @@ Not yet released.
 * Improve source plural check accuracy.
 * Update list of user interface languages in Docker.
 * Improved error messages when creating pull requests.
+* Fixed creating pull requests on Pagure.
 
 Weblate 4.6
 -----------